+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 
 ### Added
 
+- [EV-1909](https://eit-oxford.atlassian.net/browse/EV-1909): Added calls to log mapping csv as downloaded in portal
 - Minimum upload chunk size is increased to 10MB
 - Support for the SARS-CoV-2 pipeline.